LayoutTests/http/tests/media/video-canplaythrough-webm-expected.txt


Test that the canplaythrough event is fired after starting to load a webm resource with video with audio.

EVENT(loadstart)
EVENT(canplaythrough)
END OF TEST
